Migrating ThemeChooser code
In a basic replication of the recent work I did on the NumberPreference class (510a02004967a464c7432d7da4939af26b84b8e6), I migrated the ThemeChooser class to the AndroidX libraries. This involved splitting the class in two, and adding them to the onDisplayPreferenceDialog method. To prevent repeating code, I also redesigned the flow of the onDisplayPreferenceDialog method, splitting off the non-preference-specific code into a separate block.