Adjusted header size
For some reason, on what I would consider 'normal' screens (5.2"), the font size of the section headers on the reference and question selection screens would double in size when compared to smaller screens (4"). One might assume it's the normal operation to maintain some aspect of readability or something, except it's just those headers built by the two buildHeader methods. Everything else maintains its proportional size, it's only the headers that balloon like that. So I'm specifying that it should be smaller on those screens, so it maintains a consistent size. There's probably a better solution, but this one works, so I'm keeping it.