Added all Kyōiku Kanji
The remaining Kyōiku Kanji, which is taught to all Japanese school children, that hasn't already been included in the app, have been added. I'm already covering grades one and two. So this covers grades three to six. Had to make a few slight adjustments to the xsd to support it.