Adjusted chart label fonts
The axis labels on the LogDetailView chart had two problems: The typeface wasn't changing with the app theme, and instead was perpetually sans-serifed; and they were a bit too small to be properly legible. So I set the font to change with the theme, and increased the font size. I also had to increase the padding at the bottom with the setExtraBottomOffset method, because otherwise the slightly larger x-axis labels were being cut off by the bottom of the View object. Took me a while to find that method as well.