Fixed issue with tiny LogDetailItems
Was having a problem where, on screen rotation, the size of the items on the LogDetailView screen would shrink to almost nothing. I quickly determined this was caused by the correctFontSize method, and likely because at some point it attempts to run the onDraw method when there is zero space available to it. So I fixed this by basically having the assigned font size recorded in a special variable, and every time the correctFontSize method is called, rather than starting from the current font size, whatever it is, it starts from the original font size. This worked, except that the rotation was still causing the height of the objects to shrink dramatically. I fixed this by having the object height stored in a special variable, and never changing in response to the correctFontSize method. This is mainly done by only having it set once by the initial setFontSize method, and never any time else.