Fixed infinite loop on LogView
In a similar issue to the one I was experiencing on the LogDetailView screen (89a3aa57119c3192345388d193eb99d5b2ddad5f), I've been getting caught in infinite loops on LogView, and it was caused by the correctFontSize method in DailyLogItem. And it seems it was for similar reasons, because it was attempting to run the onDraw method when there is very little horizontal space. Only difference was the design of the correctFontSize method only changed the font size of larger items, and not the date label at the beginning. This was preventing the shrinkage error, but instead was causing an infinite loop, because if the available width was smaller than the width of the date label, the correctFontSize method could never shrink the text to fit in the available space, since even a font size of zero would not be small enough. So I threw in a hook to stop the code from running if the available width was that small. This prevents the infinite loop, but there's still the possibility one could experience the shrinkage error as well. However I didn't so I'll forget about it for now.