Adjust label size on small screens
After the recent change to fonts on the chart's axis labels (fdfc69fca21564d4daccea493bafab30d4bf27e4), which was made because the text on my testing device was too small, I discovered that on small-screen devices the text was significantly smaller. So I made a few adjustments to fix this. To start I moved the font size I was using into the dimens resource file, and assigned a significantly bigger number to small-screen devices. I also increased the value sent to the setExtraBottomOffset method, because the text was being cut off on small-screen devices. This was all done for small-screen devices, basically.