Redesigned landscape view
The layout for the MainQuiz's landscape mode has been adjusted in various ways. The size of the question box is now fixed at something more appropriate, and not always taking up 50 per cent of the screen. Also the descriptive instruction is now anchored to the bottom of the screen instead of floating an arbitrary distance away. I also adjusted the positioning of the AnswerFrame object and its contents so it's anchored to the top of the screen, and centered across the remainder of the screen that's not taken up by the question box. I mainly did all this because I noticed one moment where the multiple choice answer buttons were overflowing the screen just slightly in landscape mode. This change gives the AnswerFrame more room to fit the buttons.