Added notes recording theme deficiencies
Themes on pre-Oreo (API <26) don't work fully. Each section of APIs has its own deficiencies which I've been unable to crack thus far. So I went over each of them, recorded each of the problems I have yet to solve, and added them as a note to each of the style resource files. Mostly it's font issues, but there's also the problem with backgrounds on pre-Lollipop (API <21). Anyway, this is partly to remind myself where the issues are so I don't freak out that something broke, but also to remind myself to find a fix in the near future.