Fixed issue with font preview
I realized what was wrong with my use of fonts in the theme chooser font previews, which is an issue I noticed in the previous commit (6425dff580bbede3da37e2275c103f018a4ab084). I was supposed to use different font attribute ids for pre and post-Oreo (API 26). So I just extended the getDefaultThemeFont method so it could accept arbitrary themes, and not just the current one, much like most of the other ThemeManager methods. And then, by implementing that method in ThemeSelectionItem class, the issue was solved. I also removed the related notes from the style resource files.