Added gap on LogDetailItem
The text for the question and the ratio on LogDetailItem was running together in some cases. Specifically whenever it was a really long vocabulary word. So I made a quick tweak that forced an 8dp gap between these two items. This is specifically designed to work when used in conjunction with the auto size feature. If it's disabled, there's nothing I can really do. But I don't disable it, so it's not an issue.