Altered icon for dev version
To further differentiate the icons of the release and development versions of the app, I changed the background colour of the icon for the development version to red. This will only work on Oreo and later (API >= 26), but that is the version I'm running on my phone, which is why I'm making this change. It's so I can have both versions installed without confusing them.