Moved vocabulary titles
It never really made sense for the titles of the vocabulary sets to be stored as string resources. Mainly because that only really comes in handy for translation purposes, and any translators should also be translating the original vocabulary.xml files. So might as well just store the titles directly in the xml files, rather than through string references. It'll make both development and translation slightly easier. I plan to do the same thing to the kanji set titles as well.