Fixed crash caused by theme feature
Turns out, when migrating from an earlier version of The Kana Quiz, the new theme feature doesn't properly set the default theme out of the gate. This causes a crash on startup in the getThemeFont method. So I changed the getThemeId method, which manages how the theme is loaded, so if there's a glitch, it reverts to the default theme: 'Cherry Blossom Tree'. This will prevent ridiculous crashes on start up.