Added copyright messages
Decided to add the copyright boilerplate I set up in the last commit (6b8962cd37e271842020f0bd2a0d5b884a560b9e) to every source code file in the project. I also rolled back the dates on the copyright messages so they matched up with the last significant edit as recorded in the git history. I don't know if that's wrong in some way, but I'm doing it anyway. I also added the name of the spanish/catalan translator to the copyright boilerplate of the files she contributed to, mainly because it felt wrong not to.