Removed licence boilerplate from string resources
The licence boilerplate I placed in the string resources file was not being used, and the only reason I added it was because I wanted to include it in the About screen, but couldn't find a way to do so. Also, I wasn't sure what I was supposed to do with that boilerplate in the first place. Now that I figured that part out, by placing it at the top of every source code file (e6f57f87860c4d61d06c32f09336457387dea3df), I decided to remove it from string resources. I'm never going to add it to the About screen, it was a dumb idea to begin with.