Created details dialog box
Tapping on an item on the reference screen will now pop up a small dialog box that is intended to display the full details of the specific reference item. Still a work in progress, which is why it currently displays a bit of dummy data. But it's a good start.