Added WordQuestion details to ReferenceDetail
Wrote the code to fetch the material for the ReferenceDetail class in the WordQuestion class. I had some rendering problems, so to fix it, I also altered the layout files to 'wrap_content' for item size. Also tweaked the padding, and removed a redundant 'text' attribute. Because the text is set by the java code anyway.