Modified KanaQuestion detail output
For efficiency, rather than repeat the same romaji for each system it's referenced to, I altered the code so it instead just lists the romaji once, and all the systems referenced next to it. Doing so meant iterating through the altRomaji, and rearranging the data temporarily so I can properly output it. But the end result certainly looks nicer, and is easier to read.