New getters and setters for string sets
Found out the MultiSelectListPreference class stores the info as a string set, so I need to create new getters for that, and decided to create setters to go along with them, even though I'm unlikely to use the setters. Like all the getters and setters in OptionsControl, I created one of each that accepts a resource id as the key, and another of each that accepts a string as the key.