Created new KunYomiQuestion and OnYomiQuestion classes
Created two new classes for kun'yomi and on'yomi questions that are extended from the KanjiQuestion class. From there, I only needed to override the methods that dealt directly with the answers to the questions. This does result in a bit of overhead, since all four KanjiQuestion properties are duplicated for each of these questions, but this'll work for now.