Made database keys unique for new classes
The getDatabaseKey method in the KunYomiQuestion and OnYomiQuestion classes has been altered slightly so the result'll be unique, and not a duplicate of the parent KanjiQuestion key. My method of doing this is to just append a 'k' or 'o' to the kanji, which is a bit crude, but it'll do for now. I hope to improve this in the near future, but it'll likely involve a partial database redesign, which is a bit of an ask, so I'm putting it off for now.