Edited the QuestionBank class to manage new question classes
The two new question classes (KunYomiQuestion and OnYomiQuestion) are now properly handled by the QuestionBank class. I mainly did this by just replicating the code used for the vocab questions classes (KanjiQuestion and WordQuestion), because I want the multiple choice answer list to be unique for these questions, and not have it mixed with the simpler KanaQuestion answers.