Fixed tiny error in getPossibleKanaAnswers
There's a line of code in the getPossibleKanaAnswers method that's designed to make answers originating from diacritic questions pop up more frequently on diacritic questions, and answers originating from digraph questions pop up more frequently on digraph questions. But I was looking over this code while working on a separate branch (kanji-sound-questions), and I noticed the 'if' statement that actually checks this stuff out didn't look right. It was checking if the 'special list' contained the current question's database key, which didn't really make sense. I, for a minute, thought that maybe it was something that did make sense at the time, and was just a bizarre side-effect of how I designed it, somehow. But no, it was an actual error. So I corrected it. The only reason I didn't notice it sooner was because the effect of this section of code isn't really obvious.