Getting the new question classes to load
Altered the QuestionBank class to load the new question classes. I wanted to have them load through the QuestionManagement class, but I realized there was no easy way to do that without redesigning how the questions are loaded through XML, or unloading and looping through the KanjiQuestion groups to fetch the new questions. And the unloading and looping is done in the QuestionBank class anyway, so I decided I'd put the relevant code there, in the QuestionBank class. My first issue was how I was going to implement this. I couldn't just squeeze in a simple line in the 'for' loop that goes through the questions, because the new questions have to go through the same process as all the others. So I determined the simplest approach was to move the 'for' loop code into a new method so I can have the KanjiQuestions run through it multiple times for each question type. I can say with certainty that this is the best approach.