Added new questions to MainQuiz
Decided to add some code in MainQuiz to handle the new question classes so the proper prompts would be displayed. However, I ran into an issue. You see, the only way to determine the type of question being asked was through the isCurrentQuestionVocab and isCurrentQuestionKanjiSound methods. This was fine before now, because there were only two possible and relevant states, either vocab or kana. But with the new questions, I now needed to consider this new state. The isCurrentQuestionKanjiSound method was intended to fix that, but when looking at it from the MainQuiz class, I also needed to differentiate between questions asking for the kun'yomi and the on'yomi, because those are two different messages I need to output to the user. I considered making a new boolean method for this purpose, but I realized this was getting out of hand. So instead, I replaced the isCurrentQuestionVocab and isCurrentQuestionKanjiSound methods with a new getCurrentQuestionType method, along with an enum to support it. Using this new design makes the output clearer and more functional. Especially since there was no real reason to have two methods for this function, and the fact that the whole scheme was actually left a little ambiguous, what with both of them returning false signifying a KanaQuestion. Now everything's much clearer. I also added the new messages to the String resource file.