Added default value to kanji question preference
The new kanji question type preference really needed some type of default setting. I tried just placing the single option I wanted selected as the default, but that was causing crashes on startup. I presume this is because it was expecting an array of strings, and not a single string. So I created a new string array resource, containing only the single item I wanted selected, and assigned that as the default.