Fixed false positive lint inspection
The lint inspection was throwing up a false positive in the setTheme method, because it does not recognise the 'if' statement restricting the code to post-Lollipop devices (API >= 21) if it contains additional conditions linked by 'and' operators. This was causing build errors on various CIs for a while, but I was too aloof to pay attention at that time. So I finally fixed it by reorganizing the 'if' statement so the additional conditions are instead in a separate 'if' statement. There's a good chance I did this before, but the change I made disabling the FontProvider on Pie (API 28) (4d5b3babf21477609c7331c6172e804c499434d2) probably messed it up and I simply didn't notice. So, I also added a note reminding me this is why the two 'if' statements are structured this way, so I don't mess it up again. Hopefully, this fully corrects the issue.