Added new preference to getCurrentPrefRecord
The getCurrentPrefRecord method is designed to record exactly which preferences are currently set, so if (and only if) they change, the question bank gets rebuilt. But it's meant to only store the preferences that are relevant to question selection, and the kanji question type preference is one of those preferences. So I added it to the record. Thankfully I don't have to do anything beyond this, because the system is designed to merely check if anything changed, and not what changed. Any changes whatsoever cause the entire question bank to be refreshed. Simple.