Replaced deprecated firebase-plugins
Learned that the firebase-plugins library has been deprecated. So I replaced it with the perf-plugin library, since that's what the buildRelease command's output recommended. Though it did suggest using v1.2.1, but it's probably because the message is old, since v1.3.1 was available, so I'm using that. Don't know if there is anything else that needs to be imported due to this change, but this'll do for now.