Added date to dev version name
There are moments when I'm running the dev version of the app, when I'm not sure which version I actually installed on my phone. This change simply puts the date and time in the app's displayed version name, allowing me to easily tell when the build was actually made. So, if it was made well before I think it should've been, then I know it's the wrong version. It's an easy fix, and won't affect any of the standard release versions (release and pure) of the app.