Fixed crash on update from older version
The app was crashing on start up whenever the app was upgraded from a version that did not have the kanji question preference. This was because the startup sequence involved the QuestionManagement class initializing, and recording current preferences. But oddly, at that stage, the default settings weren't being set, and the act of retrieving the kanji question type preference would return a null value. As a result, examining this would throw a NullPointerException. I originally started putting in a hook that would replace the null with an empty TreeSet, but knew the ideal solution would be to ensure the default settings would be set properly. Looking in the initialize method in OptionsControl, I noticed the setting of default values was the last line of code in the method. Tried testing this to check if placing it earlier actually made a difference, I thought it did at first, but then realized that trying to reproduce the original error wasn't working. But that was because I thought it would also trigger on a fresh install. Also the crash wouldn't trigger if the preferences weren't altered in any way in the app version we're upgrading from. So, after finding a way to reliably trigger the crash, so I could check if the issue has been resolved, I found my fix wasn't working. Examining the code a bit further, I noticed the final parameter, called 'readAgain', was set to false. Looking at the documentation, I realized that the readAgain parameter was preventing the defaults from being set if it's not on a fresh install. Changing that to 'true' fixes the issue. It does mean the method runs on start up every time. But it will not override the user's settings, I can be sure of that. This also reminds me why I don't use Kotlin. That NullPointerException was actually very valuable. Anyone says it's better not there, is wrong.