Expanded QuestionType enum
I had an idea to use the QuestionType enum for purposes outside the QuestionBank. But first, it needs to be expanded somewhat. So I added KANJI as a separate item, and expanded VOCAB to VOCABULARY for the sake of clarity. Had to also expand certain bits of code to deal with this change, so there are no material changes to the overall operation of the app.