Split off QuestionType enum
I split off the QuestionType enum into a separate file. This is mainly because I want to use it in other things, as I mentioned last time (9a5de897176e230a4a14918da9edc1ab84ae1222). Because of that, it makes little sense to keep it confined within the QuestionBank class. I also transferred the code from getCurrentQuestionType into a static method in the new QuestionType enum. Also, I decided I'd use that method, for the sake of efficiency, in the addQuestion method, so it no longer has to call the getClass method several times over.