Added 'type' field to database
Added a new 'type' field to the QuestionRecord and IncorrectAnswerRecord. Mainly to differentiate between basic kanji questions, and the new kun'yomi and on'yomi questions in the database. This also would work to clarify the type of question used in other cases, because it's been slightly ambiguous up until now. As part of this, I had to add a 'type' parameter to many methods in the DAO. For the sake of consistency, I had them each follow the 'question' parameter.