Wrote migration code
Wrote the necessary code for migrating users to the new database design without losing data. Unlike previous migrations, I could not use a single command to manage the transfer, simply because part of this migration involves actually sorting which 'type' category each question belongs in from just looking at the question text. So I actually have to loop through the items and analyse each one individually in Java code to determine which 'type' the question belongs to. Because of this, I had to look back at some old code (30c55cc1394a3a3977a9ba05b571e1fe8cfed5d2) from back when the migrations were written to loop through the items individually, before I rewrote them to handle all that in SQL code (a960ccaef37e9097a6ebea29b3a8cbfa2b50e88d). I copied some of the old code and tweaked it to fit the new migration. As part of this, I also wrote a new method that is designed exclusively to determine which type of question we're looking at from just the text. I needed to do this because it's being used for two different tables in two distinct sections of the migration, and I don't want repeating code. The whole thing still needs to be tested, but it should work.