Fixed setting of 'type' field in migration
Seems the migration code was storing a solid 0 in the type parameter, which is actually invalid. I was thrown because it's intended to store a QuestionType enum, but I had to write a converter to convert it to char, but the Room library is storing it as an integer. I assumed that meant the 'execSQL' method would accept a char as a parameter, and do the necessary conversion itself. Turns out that's not the case and it actually doesn't accept chars at all. It must've been panicking and storing it as a null. I figured this out because the similar 'query' method actually throws exceptions in that case, but 'execSQL' does not. So I fixed it by storing the char in an int variable, which automatically converts it to the value I need. With this, I can say the migration code is finally working.