Added new migration test
Rather than test all the questions to ensure they're all sorted properly in the normal migration test, I decided to create a new test exclusively for this function. Since I'm able to design the new test to have a single function, it's much simpler, and I can remove a lot of the items from the normal test, since they're handled in the new test. I still have to add vocabulary questions, but that'll be a lot easier now.