Cleaned up DatabaseKey handling
In order to ensure the kun'yomi and on'yomi questions both had unique database keys, I had to override the getDatabaseKey methods, and add a one-character suffix to the key, so it wouldn't share it's key with its equivalent KanjiQuestion object (dd5516d062ed1ca66ba161d78dc7b7fa920e56e2). But since I redesigned the database, that's no longer an issue. So I reverted that change. But as a consequence, I had to make a few more changes, since there are other places I'm using the getDatabaseKey method to generate a unique identifier. For some of them, like the current question backup that's used to prevent the current question from changing unnecessarily, I just decided to store the type alongside it. However for others, like the previous question record and the answer list cache, it became more complex, since those are structures that can only handle a single item as the determining key. So I just decided to add a single char as a prefix, which is derived from the QuestionType. Surprisingly, that toCharFromType method came in handy outside the database.