Added exceptions to QuestionType
Previously, the getQuestionType method would return a null if it was fed a class that was not derived from the Question class. I realized that probably wasn't the proper way to handle this. So now I'm having it throw an exception. Because if this happened, something has gone horribly wrong. I also did the same thing to the converters in LogTypeConversion.