Updated libraries
After going on hiatus for nearly nine months now, a large number of out-of-date libraries was inevitable. I'm now running the latest version of Android Studio (v3.6.3), which initiated a few changes on its own, and I updated all the build scripts to use the newest libraries as well as updating Gradle to version 5.6.4.