Redesigned layout of multi-character question reference
Expanded the alterations made to the layout of the ReferenceDetail box for vocab questions (b4764cce9f979c64c1cd458f097873034b4a251c) to all multi-character questions. Doing this because I ran into a few problems on smaller screens, with the digraphs filling up the whole width of the box, and the reference information rendering off-screen, and causing the wordwrap function to expand the box to its maximum height. Anyway, this should fix it.