Updated style guide
Changed the style check that checks which side of a comparison a constant is on, so constants are always on the right. It's what I already do, just on instinct, and looks more appropriate. This check seems to have been added to Android Studio only a couple years ago, and I guess the only reason I didn't notice sooner was my lack of activity. Nonetheless, the change has been made. As a consequence of this update, it seems Android Studio also removed a bunch of additional checks from the style guide file. I'm guessing this is because these checks are no longer implemented in Android Studio.