Fixed numeric issues
My attention was drawn to a few issues involving floating point numbers, and integer division. Seems in a few cases, I divided two integers without ensuring that the actual division would be using floating point numbers, when that's what I wanted. Without this change, all those calculations would just drop everything after the decimal, when I want to include that in the subsequent calculations. This was caught by the inspection system. I presume this check is a recent addition.