Migrating to ViewPager2
Since ViewPager has been deprecated, I decided to migrate all uses of it to ViewPager2. This is a bit of a complicated process, it turns out. For now, I'm just following the instructions I found online (youtu.be/lAP6cz1HSzA). But there's still a few problems to fix, which is why I started by creating a new branch. Right now, the reference screen is crashing, and I believe it has something to do with the fact that we have a ViewPager2 within a ViewPager2. I'll fix it at some point, for now I thought it would be a good idea to commit the first step, despite the fact that it's incomplete, at least it's started.