Fixed NullPointerException
With the migration to ViewPager2, I ran into a problem with the Reference screen. I was getting NullPointerExceptions every time it opened. I quickly traced the problem to the 'container' object (or an object within it) being 'null' when I attempt to fetch the context from it when generating the reference pages. I don't know why exactly this happens, or why it didn't happen with the old ViewPager library. I thought I'd have to do some restructuring or rebuilding to solve this issue, until I tried fetching the context from the 'inflater' object instead. That worked without a hitch. So I just made the simple change across the two relevant files. Everything works now.