Created new methods for OptionsControl
Created a few new methods in the OptionsControl class for the purposes of implementing finer question selection. The getQuestionSetBool and setQuestionSetBool methods are, for now, identical to the original getBoolean and setBoolean methods, but will be altered eventually, since I plan to handle the question preferences very differently to how they're handled now. This is just to set the groundwork. There are also methods to check if certain preferences exist and to delete them, which will become useful. I also made a minor change to the getBoolean method, removing the special conditional check to handle the default question settings, since question preferences are handled in a different method now.