Created QuestionSelectionDetail view
Created a simple dialog box for accessing QuestionSelectionDetails. There's still quite a bit of work to be done, but this lays the groundwork. I can't even test it though, until I find a way to pass it the list of questions within the question sets. But, I'm pretty sure it will work fine.