Finished work on QuestionSelectionDetail
Got QuestionSelectionDetail to display the specific questions for selection, by passing an array of the individual questions from the QuestionManagement class when the QuestionSelection screen is generated. I also added the necessary code to change the checked status of the parent QuestionSelectionItem object if all the individual questions are either selected or deselected.