Gave QuestionSelectionItem a neutral setting
QuestionSelectionItem now has a neutral setting, which is set when the QuestionSelectionDetail box has some, but not all, items selected. I implemented it by setting the checkbox object to be invisible. As part of this, I also made a couple changes that ensure when a question set is either enabled or disabled, all the individual questions are no longer recorded in the options, and when individual questions are either enabled or disabled, the parent question set is no longer recorded in the options. This simplifies a lot of things and allows the neutral setting of the QuestionSelectionItem to be easily recorded and checked.