Redid OnCheckedChangeListener
Changed the OnCheckedChangeListener in QuestionSelectionDetail to be more thorough. Mainly for the sake of consistency, I set it up so all the necessary preference changes are saved every time any change is made, including the wiping of the parent settings. As part of this, I also added something new. I had the detail box save the setting of every single question at the same time it wipes parent settings, since there's a good chance only one question would be selected or deselected and those that aren't touched by the user won't be saved otherwise, which would result in some real problems if I'm honest. I also removed some redundant code from the onDismiss method. I'm handling the nullify and option delete tasks in the updatePref method, and the QuestionSelectionItem will handle setting the question set preference through the OnCheckedChangeListener. Therefore, those bits of code are redundant.