Rebuilt the buildQuestionBank method
I rebuilt the method for building the QuestionBank, so it now looks at the preferences of individual questions. I had to completely redesign and rethink how the method worked. It now looks at the list of questions directly, rather than going through the getQuestionSet method. I also added a new method to the QuestionBank class, to add a single question to the bank, because that's a thing I'm doing now. To avoid repeating code, I also had the method for adding multiple questions call the method for adding a single question. Since the core operation is basically the same.