Built method for setting default question selection
Wrote a new method for setting the default options for question selection. This is necessary for the QuestionSelectionItem's neutral state to work properly. My first assumption was that I could loop through the list of prefIds immediately after the XML file is parsed, check if the setting exists, and set it to its default if it's not. But this proved problematic almost immediately because when individual questions are selected, the root preference is cleared, which will result in a false positive for the default settings checker. So I had to add in a check for the individual questions as well. As part of this I was able to remove the default check from the getQuestionSetBool method, since it's no longer necessary, and I'm pretty sure wasn't working as intended anymore anyway. I think there are still ways to streamline this method, but this'll do for now.