Expanded prefRecord
Expanded the prefRecord to store the settings for individual questions. This ended up being way more complicated than I originally intended. To start the whole thing, I knew I needed to use a 2D array. I created a new method to generate the prefRecord for the individual sets to eliminate code redundancy. It became significantly more complicated by the fact that one can't just cast an array of Boolean objects to boolean primitives, which is why I had to loop through the array of one to convert it to the other. That wasn't too difficult, just annoying. But the first major problem began in comparing the prefRecords. Turns out Arrays.equals method doesn't really work with 2D arrays, which is disappointing, so I had to write a new method specifically for comparing the two. After writing all this though, I noticed some inefficiencies which need to be dealt with. But this works, so I'm leaving it for now.