Streamlined getPrefRecord method
The method calls within the getPrefRecord method were causing some task redundancy. The getPref method checks all related preferences to give a result, which is what I'm doing anyway through the getAllPrefs method to get the full preference data, and it was being done separately because it's a separate method call. So the same thing was being done twice. I streamlined it down to a single method call that it turns out already does what I want it to do. As a result, I noticed a significant speed increase when transitioning to and from the QuestionSelection screen. So I'd say this is a job well done.