Disabled user input on viewPager
Before migrating to ViewPager2, the gestures on the ReferenceScreen would affect the subsection pager before they would affect the main pager. I want that behaviour to continue on ViewPager2, but it seems it's not doing that. Instead, it's prioritizing the main pager. I get the feeling this is going to get complicated. For now, I disabled user input on the main pager, so swipe gestures would only affect the subsection pager, because I feel that's more useful. I plan to make further changes in this area, but in the meantime, this'll do.