Replaced comparePrefRecord
I found a built-in method that can compare multi-dimensional arrays, like the Arrays.equals method, called deepEquals. Yesterday, I wrote a method that does this for the prefRecord comparison, and now I realize how useless it is, so I decided to replace it with deepEquals. Basically, this is a partial reversion of an earlier commit (7abe389b71e084c259ba0fd251ddadbe6b404c58).