Redesigned entire prefRecord
Completely redesigned the prefRecord system to be stored as an array of integers, as opposed to booleans. There's an advantage to this new design. An integer is 32 bits, which means functionally, 32 booleans can be encoded in a single integer. And because of how booleans and integers are stored, from what I know, and what I can tell, I estimate that this will reduce the amount of space used to store the prefRecord down to 1/8th of what it was previously. I needed to use an array of integers, because there are a lot more than 32 booleans I need to store. From what I can tell, this not only improved memory usage, but speed as well. While the code is a bit lengthy, and there might still be room for improvement here, it is the most efficient, and simplest, way of storing this volume of information. I had this idea for a while now, but given the substantial increase in information I need to store now, with the addition of individual question preferences, it has never been more necessary.