Improved how diacritics are stored in prefRecord
Instead of recording the setting of the diacritics option as a separate boolean, I decided to signify that setting by whether or not the diacritic questions are set to true. Since I'm now storing individual questions, it feels appropriate. It also means changing the diacritics setting will not be recognised as a prefRecord change if no diacritic questions are selected in the first place.