Fixed ConcurrentModificationException issue
I was occasionally getting a ConcurrentModificationException error when testing, which I found difficult to solve, because I didn't know what it was. Because it wasn't consistent, I assumed it was an issue with multithreading. However after looking up the error, I discovered it was actually caused by editing a collection while iterating over it. Which makes sense, but I never considered. It was such an occasional error that I never bothered to look into it. Good news is, once I knew the cause, it was an easy fix. Instead of editing the collection, I create a new collection to replace the first one. I do believe this sufficiently solves the problem.