Redesigned inefficient loops
There were some inefficient loops that would loop through the entire array of question sets looking for specific sets. This is mainly because I kinda forgot that I could just grab the specific numbered sets. I think I just got used to the idea of looping. Anyway, I redesigned them to fetch the specific question sets rather than look through all of them and filter out the one I want with if statements. This is important, because the way I was doing it before was ridiculously inefficient, with the app having to loop through the whole thing multiple times, unnecessarily.